requirements.txt
fastapi
uvicorn
langchain
openai
pinecone-client
python-docx
pdfminer.six
python-multipart
streamlit
wordcloud
fpdf